refactor(storage): Remove unused SharedDatabase from storage crate
The storage crate's SharedDatabase (parking_lot::RwLock-based) was never
used - the server defines its own SharedDatabase as Arc<tokio::sync::RwLock>
for async compatibility. This removes the dead code to eliminate the name
collision and confusion between the two types.

Closes #4258
