+++ crates/vibesql-executor/Cargo.toml
@@ -59,6 +59,7 @@ vibesql-parser = { version = "0.1.3", path = "../vibesql-parser" }
 # NOTE: vibesql-server dev-dependency removed to break circular dependency.
 # Server benchmarks (sysbench_server, tpcc_server, tpch_server) are now in vibesql-server crate.
 criterion = { version = "0.8", features = ["html_reports"] }
+parking_lot = "0.12"  # For concurrent_throughput benchmark
 rand = "0.9"
 rand_chacha = "0.9"
 rust_decimal = "1.33"
